ported classes which exists on CPP  version BusTub only (data structure, skelton of func, call of skelton func)

+++ container/hash/hash_table_page_test.go
@@ -17,7 +17,7 @@ import (
 func TestHashTableHeaderPage(t *testing.T) {
 	diskManager := disk.NewDiskManagerImpl("test.db")
 	//bpm := buffer.NewBufferPoolManager(diskManager, buffer.NewClockReplacer(5))
-	bpm := buffer.NewBufferPoolManager(10, diskManager)
+	bpm := buffer.NewBufferPoolManager(uint32(10), diskManager)
 
 	newPage := bpm.NewPage()
 	newPageData := newPage.Data()
